+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_commands.py
@@ -13,6 +13,8 @@ from azure.cli.core.azclierror import ValidationError
 from azure.cli.testsdk.scenario_tests import AllowLargeResponse, live_only
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer, JMESPathCheck)
 from msrestazure.tools import parse_resource_id
+from _dev_service_utils import addon_provisioningState_check
+from _constants import DEV_SERVICE_LIST
 
 from azext_containerapp.tests.latest.common import (write_test_file, clean_up_test_file)
 from .common import TEST_LOCATION
@@ -794,6 +796,7 @@ class ContainerappServiceBindingTests(ScenarioTest):
         kafka_ca_name = 'kafka'
         mariadb_ca_name = 'mariadb'
         qdrant_ca_name = "qdrant"
+<<<<<<< HEAD
         ADDON_LIST = ["redis", "postgres", "kafka", "mariadb", "qdrant"]
         create_containerapp_env(self, env_name, resource_group)
 
@@ -820,16 +823,48 @@ class ContainerappServiceBindingTests(ScenarioTest):
             resource_group, ca_name, env_name, image), checks=[
             JMESPathCheck('properties.template.serviceBinds[0].name', "postgres_binding"),
             JMESPathCheck('properties.template.serviceBinds[1].name', "redis")
+=======
+
+        create_containerapp_env(self, env_name, resource_group)
+
+        self.cmd('containerapp addon redis create -g {} -n {} --environment {}'.format(
+            resource_group, redis_ca_name, env_name))
+
+        self.cmd('containerapp addon postgres create -g {} -n {} --environment {}'.format(
+            resource_group, postgres_ca_name, env_name))
+
+        self.cmd('containerapp addon kafka create -g {} -n {} --environment {}'.format(
+            resource_group, kafka_ca_name, env_name))
+
+        self.cmd('containerapp addon mariadb create -g {} -n {} --environment {}'.format(
+            resource_group, mariadb_ca_name, env_name))
+
+        self.cmd('containerapp addon qdrant create -g {} -n {} --environment {}'.format(
+            resource_group, qdrant_ca_name, env_name))
+
+        for addon in DEV_SERVICE_LIST:
+            addon_provisioningState_check(self, addon_name=addon, resource_group=resource_group)
+
+        self.cmd('containerapp create -g {} -n {} --environment {} --image {} --bind postgres:postgres_binding redis'.format(
+            resource_group, ca_name, env_name, image), checks=[
+            JMESPathCheck('properties.template.addonBinds[0].name', "postgres_binding"),
+            JMESPathCheck('properties.template.addonBinds[1].name', "redis")
+>>>>>>> 9d479eb49 (add)
         ])
 
         self.cmd('containerapp update -g {} -n {} --unbind postgres_binding'.format(
             resource_group, ca_name, image), checks=[
+<<<<<<< HEAD
             JMESPathCheck('properties.template.serviceBinds[0].name', "redis"),
+=======
+            JMESPathCheck('properties.template.addonBinds[0].name', "redis"),
+>>>>>>> 9d479eb49 (add)
         ])
 
         self.cmd('containerapp update -g {} -n {} --bind postgres:postgres_binding kafka mariadb qdrant'.format(
             resource_group, ca_name, image), checks=[
             JMESPathCheck("properties.provisioningState", "Succeeded"),
+<<<<<<< HEAD
             JMESPathCheck('properties.template.serviceBinds[0].name', "redis"),
             JMESPathCheck('properties.template.serviceBinds[1].name', "postgres_binding"),
             JMESPathCheck('properties.template.serviceBinds[2].name', "kafka"),
@@ -850,11 +885,37 @@ class ContainerappServiceBindingTests(ScenarioTest):
             resource_group, mariadb_ca_name, env_name))
 
         self.cmd('containerapp add-on qdrant delete -g {} -n {} --yes'.format(
+=======
+            JMESPathCheck('properties.template.addonBinds[0].name', "redis"),
+            JMESPathCheck('properties.template.addonBinds[1].name', "postgres_binding"),
+            JMESPathCheck('properties.template.addonBinds[2].name', "kafka"),
+            JMESPathCheck('properties.template.addonBinds[3].name', "mariadb"),
+            JMESPathCheck('properties.template.addonBinds[4].name', "qdrant")
+        ])
+
+        self.cmd('containerapp addon postgres delete -g {} -n {} --yes'.format(
+            resource_group, postgres_ca_name, env_name))
+
+        self.cmd('containerapp addon redis delete -g {} -n {} --yes'.format(
+            resource_group, redis_ca_name, env_name))
+
+        self.cmd('containerapp addon kafka delete -g {} -n {} --yes'.format(
+            resource_group, kafka_ca_name, env_name))
+
+        self.cmd('containerapp addon mariadb delete -g {} -n {} --yes'.format(
+            resource_group, mariadb_ca_name, env_name))
+
+        self.cmd('containerapp addon qdrant delete -g {} -n {} --yes'.format(
+>>>>>>> 9d479eb49 (add)
             resource_group, qdrant_ca_name, env_name))
 
         self.cmd(f'containerapp delete -g {resource_group} -n {ca_name} --yes')
 
+<<<<<<< HEAD
         self.cmd('containerapp add-on list -g {} --environment {}'.format(resource_group, env_name), checks=[
+=======
+        self.cmd('containerapp addon list -g {} --environment {}'.format(resource_group, env_name), checks=[
+>>>>>>> 9d479eb49 (add)
             JMESPathCheck('length(@)', 0),
         ])