+++ src/containerapp/HISTORY.rst
@@ -4,7 +4,11 @@ Release History
 ===============
 upcoming
 ++++++
+<<<<<<< HEAD
 * 'az containerapp add-on': support for az containerapp add-on commands; deprecation of az containerapp service commands
+=======
+* 'az containerapp addon': support deprecation of service command in favor of addon.
+>>>>>>> 9d479eb49 (add)
 * 'az containerapp env create': Support --enable-dedicated-gpu
 * 'az containerapp job create': fix problem of parsing parameters minExecutions and maxExecutions from --yaml
 * 'az containerapp env dapr-component init': support initializing Dapr components and dev services for an environment